Ensure django-recurrence JS works in all browsers (fixes #27)
Chrome and Firefox appear to handle this differently:

if (!foobar) {
function foobar() {
    alert("hello");
}
}

In Chrome, foobar is always defined after that code is run.
In Firefox, foobar is only defined if it was defined before
that code is run. The easiest way to fix this is to create a
local variable, and define it if it's not defined already
(which is what the previous code attempted to do), like
this:

var foobar = foobar || function() { alert("hello"); };

This change also fixes a problem introduced in #21, where we
changed the name of the translations catalog from catalog to
django.catalog, which didn't work, since we've got no
control over the name of the variable in
django.views.i18n.javascript_catalog which stores
translation strings. I can only think that that change was
made somewhere which happened to copy that in to an object
called "django".

Finally, this change documents how to use the
internationali(s|z)ation support for django-recurrence.