Allow negative BYMONTHDAY.
RFC 2445 allows negative values and gives them meaning of a count
from the end of month; zero is not allowed and would be ignored.

This is not a very smart approach but it's a start. The monthly
form in the widget displays a hardcoded number of extra days at the end
of the calendar.