Added 4 lines missing from get_media() in forms.py
The original issue was described here:
https://github.com/django-recurrence/django-recurrence/issues/170#issuecomment-710028328

Django-recurrence widget was found to be functioning properly in the Django Admin, however when integrating the widget into a custom Model Form the buttons had no functionality when clicked. It was found that 2 Javascript tags and 2 CSS tags were missing that shouldve been loaded automatically through {% load static %} and {{ form.media }}. Looking through the source code it seems that 4 lines were missing from the get_media() method in forms.py which otherwise need to be loaded manually by anyone using the widget to get proper functionality.